requirements/requirements-testing.txt
codecov
tox
mock
pkgversion
pytest-django
pytest-cov
readme_renderer[md]
twine
setuptools
wheel